+++ Week_09/homework/dubbo-hmily-demo/transaction-api/src/main/java/com/walker/transaction/rmb/RmbAccountFreeze.java
@@ -19,5 +19,5 @@ public class RmbAccountFreeze {
     @Id
     private Integer userId;
     private Integer amount;
-    private Integer freezeType;
+    private Long freezeType;
 }
