+++ Week_09/homework/dubbo-hmily-demo/transaction-api/src/main/java/com/walker/transaction/usd/UsdAccountFreeze.java
@@ -19,5 +19,5 @@ public class UsdAccountFreeze {
     @Id
     private Integer userId;
     private Integer amount;
-    private Integer freezeType;
+    private Long freezeType;
 }
